+++ static/attraction.html
@@ -35,6 +35,6 @@
     </div>
     <div class="bottom"></div>
     <footer>COPYRIGHT © 2024 台北一日遊</footer>
-    <script type="module" src="/static/js/attraction.js"></script>
+    <script type="module" src="/static/js/attraction/index.js"></script>
   </body>
 </html>
